# openapi.yaml
openapi: 3.0.3
info:
  title: 协同管理工具 API
  version: 0.1.0
  description: |
    Spec-first 唯一契约源。
    Go 侧用 oapi-codegen 生成 server 接口与模型，前端用 openapi-typescript + openapi-fetch 生成类型化客户端；生成物不手改。
    业务规则只在后端 domain 包实现，界面反馈一律消费本契约返回的派生字段。
servers:
  - url: /api/v1
...
paths:
...
  /projects/{projectId}/blockers:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: listBlockers
      tags: [tasks]
      summary: 项目当前派生的结构化卡点（风险队列数据源；四类结构化事实读时派生）
      responses:
        "200":
          description: 卡点列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blocker"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/reminders:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: createReminder
      tags: [tasks]
      summary: 一键提醒当前待行动人（AC-11、MW-13）：目标既可以是派生卡点，也可以是尚未成卡点的等待事项（审批件、输入请求、上游任务）；通知自动带入任务、缺失输入、截止时间与下游影响
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemindRequest"
      responses:
        "204":
          description: 已提醒
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    Unauthorized:
      description: 未登录或会话已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: 请求内容未通过校验
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 当前用户无权执行该动作（动作权限按成员角色在 domain 层统一判定）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 与现有资源状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 机器可读错误码
        message:
          type: string
          description: 面向用户的错误说明
...
    RiskLevel:
      type: string
      description: 风险等级（正常／预警／高风险），见词汇表「风险等级」；读时派生值，不落库、无写入路径（KR 取「下级任务未解除卡点最高等级、超期、临期」的最大值，PRD §5.7）
      enum: [normal, warning, high_risk]
...
    BlockerKind:
      type: string
      description: 四类派生卡点（词汇表「结构化卡点」；我的工作 PRD §8.7）
      enum: [upstream_unready, task_overdue, approval_timeout, interlock]
...
    Blocker:
      type: object
      description: 结构化卡点（词汇表）；由四类结构化事实读时派生，不落库、无人工上报与手动解除，触发条件消失即消失
      required: [key, kind, kindLabel, taskId, taskName, missing, reason, level, actionOwnerIds, actionOwnerNames, since]
      properties:
        key:
          type: string
          description: 派生卡点的合成键（形如 task_overdue:42、upstream_unready:edge:17）；一键提醒按此寻址
        kind:
          $ref: "#/components/schemas/BlockerKind"
        kindLabel:
          type: string
          description: 四类卡点的中文类型名（派生字段）；行级显示消费本字段，前端不按枚举拼文案
        taskId:
          type: integer
          format: int64
        taskName:
          type: string
        missing:
          type: string
          description: 缺失的输入或条件
        reason:
          type: string
          description: 阻塞原因
        actionOwnerIds:
          type: array
          description: 当前待行动人（成果审核（或签）、互锁环内 KR 负责人可为多人）
          items:
            type: integer
            format: int64
        actionOwnerNames:
          type: array
          items:
            type: string
        level:
          $ref: "#/components/schemas/RiskLevel"
          description: 按事实严重度派生的预警或高风险（结构化卡点不使用 normal）
        levelLabel:
          type: string
          description: 风险等级显示文案（派生字段；前端不按枚举拼文案）
        since:
          type: string
          format: date-time
          description: 卡点产生时间（触发事实成立的时点）
        impactNote:
          type: string
          description: 影响范围一行说明（派生字段；沿硬前置下游任务汇总）
        canRemind:
          type: boolean
          description: 当前用户能否一键提醒（派生字段；待行动人本人不可提醒自己，访客不可）
    RemindRequest:
      type: object
      required: [targetKey]
      properties:
        targetKey:
          type: string
          maxLength: 200
          description: 提醒目标的合成键；卡点取自卡点列表或任务详情的 key，等待他人事项取自我的工作卡片的 refKey（形如 wait:<事项类型>:<事项 ID>）
...
security:
  - cookieAuth: []